error on clean using powershell
I am getting the error "A resource must be a file, an empty folder or a symlink" when going through the clean process on PowerShell.  While doing some debugging I found that the RecursiveDirectoryIterator was calling delete on the "." directory.  I have added the FilesystemIterator::SKIP_DOTS flag to get around the issue.